move UI initialize code from `main` to `view` module
